Refactor project components * Remove bin folder, leaving cli, server, and client * This fixes #315 * In .gitignore,   - add .idea   - remove bin/build * Remove bin and bin/cli.js * Remove bin/scripts/copy-cli.js * Refactor/move bin/scripts to cli/scripts * Refactor/move bin/src/index.ts to cli/src/cli.ts * Refactor/renamed client/bin/cli.js to client/bin/client.js * In .github/workflows/main.yml,   - add run of cli tests * In cli/pacakge.json   - change main and bin/mcp-inspector-cli properties to build/cli.js * In client/package.json,   - change bin/mcp-inspector-client properties to build/start.js * In pacakge.json   - change bin/mcp-inspector property to ./cli/build/cli.js   - removed bin and cli/bin from files list   - removed @modelcontextprotocol/inspector-bin dependency   - rearranged and corrected scripts
