--- bin/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@modelcontextprotocol/inspector-bin",
-  "version": "0.9.0",
-  "description": "Model Context Protocol inspector",
-  "license": "MIT",
-  "author": "Anthropic, PBC (https://anthropic.com)",
-  "homepage": "https://modelcontextprotocol.io",
-  "bugs": "https://github.com/modelcontextprotocol/inspector/issues",
-  "type": "module",
-  "bin": {
-    "mcp-inspector": "./cli.js"
-  },
-  "files": [
-    "cli.js"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "postbuild": "node scripts/make-executable.js && node scripts/copy-cli.js",
-    "test": "node scripts/cli-tests.js"
-  },
-  "dependencies": {
-    "commander": "^13.1.0",
-    "spawn-rx": "^5.1.2"
-  },
-  "devDependencies": {}
-}
