+++ cli/package.json
@@ -3,18 +3,21 @@
   "version": "0.9.0",
   "description": "CLI for the Model Context Protocol inspector",
   "license": "MIT",
-  "author": "Nicolas Barraud",
-  "homepage": "https://github.com/nbarraud",
-  "main": "build/index.js",
+  "author": "Anthropic, PBC (https://anthropic.com)",
+  "homepage": "https://modelcontextprotocol.io",
+  "bugs": "https://github.com/modelcontextprotocol/inspector/issues",
+  "main": "build/cli.js",
   "type": "module",
   "bin": {
-    "mcp-inspector-cli": "build/index.js"
+    "mcp-inspector-cli": "build/cli.js"
   },
   "files": [
     "build"
   ],
   "scripts": {
-    "build": "tsc"
+    "build": "tsc",
+    "postbuild": "node scripts/make-executable.js",
+    "test": "node scripts/cli-tests.js"
   },
   "devDependencies": {},
   "dependencies": {
