+++ cli/src/cli.ts
@@ -1,13 +1,37 @@
 #!/usr/bin/env node
+
 import { Command } from "commander";
 import fs from "node:fs";
 import path from "node:path";
 import { dirname, resolve } from "path";
 import { spawnPromise } from "spawn-rx";
 import { fileURLToPath } from "url";
+
 const __dirname = dirname(fileURLToPath(import.meta.url));
-function handleError(error) {
-  let message;
+
+type Args = {
+  command: string;
+  args: string[];
+  envArgs: Record<string, string>;
+  cli: boolean;
+};
+
+type CliOptions = {
+  e?: Record<string, string>;
+  config?: string;
+  server?: string;
+  cli?: boolean;
+};
+
+type ServerConfig = {
+  command: string;
+  args?: string[];
+  env?: Record<string, string>;
+};
+
+function handleError(error: unknown): never {
+  let message: string;
+
   if (error instanceof Error) {
     message = error.message;
   } else if (typeof error === "string") {
@@ -15,39 +39,49 @@ function handleError(error) {
   } else {
     message = "Unknown error";
   }
+
   console.error(message);
+
   process.exit(1);
 }
-function delay(ms) {
+
+function delay(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms, true));
 }
-async function runWebClient(args) {
+
+async function runWebClient(args: Args): Promise<void> {
   const inspectorServerPath = resolve(
     __dirname,
-    "..",
+    "../../",
     "server",
     "build",
     "index.js",
   );
+
   // Path to the client entry point
   const inspectorClientPath = resolve(
     __dirname,
-    "..",
+    "../../",
     "client",
     "bin",
-    "cli.js",
+    "client.js",
   );
-  const CLIENT_PORT = process.env.CLIENT_PORT ?? "6274";
-  const SERVER_PORT = process.env.SERVER_PORT ?? "6277";
+
+  const CLIENT_PORT: string = process.env.CLIENT_PORT ?? "6274";
+  const SERVER_PORT: string = process.env.SERVER_PORT ?? "6277";
+
   console.log("Starting MCP inspector...");
+
   const abort = new AbortController();
-  let cancelled = false;
+  let cancelled: boolean = false;
   process.on("SIGINT", () => {
     cancelled = true;
     abort.abort();
   });
-  let server;
-  let serverOk;
+
+  let server: ReturnType<typeof spawnPromise>;
+  let serverOk: unknown;
+
   try {
     server = spawnPromise(
       "node",
@@ -66,9 +100,11 @@ async function runWebClient(args) {
         echoOutput: true,
       },
     );
+
     // Make sure server started before starting client
     serverOk = await Promise.race([server, delay(2 * 1000)]);
   } catch (error) {}
+
   if (serverOk) {
     try {
       await spawnPromise("node", [inspectorClientPath], {
@@ -81,15 +117,20 @@ async function runWebClient(args) {
     }
   }
 }
-async function runCli(args) {
+
+async function runCli(args: Args): Promise<void> {
   const projectRoot = resolve(__dirname, "..");
-  const cliPath = resolve(projectRoot, "cli", "build", "index.js");
+  const cliPath = resolve(projectRoot, "build", "index.js");
+
   const abort = new AbortController();
+
   let cancelled = false;
+
   process.on("SIGINT", () => {
     cancelled = true;
     abort.abort();
   });
+
   try {
     await spawnPromise("node", [cliPath, args.command, ...args.args], {
       env: { ...process.env, ...args.envArgs },
@@ -102,16 +143,20 @@ async function runCli(args) {
     }
   }
 }
-function loadConfigFile(configPath, serverName) {
+
+function loadConfigFile(configPath: string, serverName: string): ServerConfig {
   try {
     const resolvedConfigPath = path.isAbsolute(configPath)
       ? configPath
       : path.resolve(process.cwd(), configPath);
+
     if (!fs.existsSync(resolvedConfigPath)) {
       throw new Error(`Config file not found: ${resolvedConfigPath}`);
     }
+
     const configContent = fs.readFileSync(resolvedConfigPath, "utf8");
     const parsedConfig = JSON.parse(configContent);
+
     if (!parsedConfig.mcpServers || !parsedConfig.mcpServers[serverName]) {
       const availableServers = Object.keys(parsedConfig.mcpServers || {}).join(
         ", ",
@@ -120,35 +165,48 @@ function loadConfigFile(configPath, serverName) {
         `Server '${serverName}' not found in config file. Available servers: ${availableServers}`,
       );
     }
+
     const serverConfig = parsedConfig.mcpServers[serverName];
+
     return serverConfig;
-  } catch (err) {
+  } catch (err: unknown) {
     if (err instanceof SyntaxError) {
       throw new Error(`Invalid JSON in config file: ${err.message}`);
     }
+
     throw err;
   }
 }
-function parseKeyValuePair(value, previous = {}) {
+
+function parseKeyValuePair(
+  value: string,
+  previous: Record<string, string> = {},
+): Record<string, string> {
   const parts = value.split("=");
   const key = parts[0];
   const val = parts.slice(1).join("=");
+
   if (val === undefined || val === "") {
     throw new Error(
       `Invalid parameter format: ${value}. Use key=value format.`,
     );
   }
-  return { ...previous, [key]: val };
+
+  return { ...previous, [key as string]: val };
 }
-function parseArgs() {
+
+function parseArgs(): Args {
   const program = new Command();
+
   const argSeparatorIndex = process.argv.indexOf("--");
   let preArgs = process.argv;
-  let postArgs = [];
+  let postArgs: string[] = [];
+
   if (argSeparatorIndex !== -1) {
     preArgs = process.argv.slice(0, argSeparatorIndex);
     postArgs = process.argv.slice(argSeparatorIndex + 1);
   }
+
   program
     .name("inspector-bin")
     .allowExcessArguments()
@@ -162,12 +220,16 @@ function parseArgs() {
     .option("--config <path>", "config file path")
     .option("--server <n>", "server name from config file")
     .option("--cli", "enable CLI mode");
+
   // Parse only the arguments before --
   program.parse(preArgs);
-  const options = program.opts();
+
+  const options = program.opts() as CliOptions;
   const remainingArgs = program.args;
+
   // Add back any arguments that came after --
   const finalArgs = [...remainingArgs, ...postArgs];
+
   // Validate that config and server are provided together
   if (
     (options.config && !options.server) ||
@@ -177,11 +239,13 @@ function parseArgs() {
       "Both --config and --server must be provided together. If you specify one, you must specify the other.",
     );
   }
+
   // If config file is specified, load and use the options from the file. We must merge the args
   // from the command line and the file together, or we will miss the method options (--method,
   // etc.)
   if (options.config && options.server) {
     const config = loadConfigFile(options.config, options.server);
+
     return {
       command: config.command,
       args: [...(config.args || []), ...finalArgs],
@@ -189,9 +253,11 @@ function parseArgs() {
       cli: options.cli || false,
     };
   }
+
   // Otherwise use command line arguments
   const command = finalArgs[0] || "";
   const args = finalArgs.slice(1);
+
   return {
     command,
     args,
@@ -199,12 +265,15 @@ function parseArgs() {
     cli: options.cli || false,
   };
 }
-async function main() {
+
+async function main(): Promise<void> {
   process.on("uncaughtException", (error) => {
     handleError(error);
   });
+
   try {
     const args = parseArgs();
+
     if (args.cli) {
       runCli(args);
     } else {
@@ -214,4 +283,5 @@ async function main() {
     handleError(error);
   }
 }
+
 main();
