+++ client/bin/client.js
